+++ backend/main.go
@@ -18,7 +18,7 @@ import (
 )
 
 const (
-	version string = "1.0.4"
+	version string = "1.0.5"
 )
 
 var kaEP = keepalive.EnforcementPolicy{