+++ backend/pb/url_service.pb.go
@@ -15,14 +15,15 @@ package protos
 
 import (
 	context "context"
+	reflect "reflect"
+	sync "sync"
+
 	proto "github.com/golang/protobuf/proto"
 	grpc "google.golang.org/grpc"
 	codes "google.golang.org/grpc/codes"
 	status "google.golang.org/grpc/status"
 	protoreflect "google.golang.org/protobuf/reflect/protoreflect"
 	protoimpl "google.golang.org/protobuf/runtime/protoimpl"
-	reflect "reflect"
-	sync "sync"
 )
 
 const (
@@ -43,6 +44,7 @@ type URL struct {
 
 	LongUrl string `protobuf:"bytes,1,opt,name=LongUrl,proto3" json:"LongUrl,omitempty"`
 	TTL     int64  `protobuf:"varint,2,opt,name=TTL,proto3" json:"TTL,omitempty"`
+	Stats   string `protobuf:"bytes,3,opt,name=Stats,proto3" json:"Stats,omitempty"`
 }
 
 func (x *URL) Reset() {
@@ -89,6 +91,13 @@ func (x *URL) GetTTL() int64 {
 		return x.TTL
 	}
 	return 0
+}
+
+func (x *URL) GetStats() string {
+	if x != nil {
+		return x.Stats
+	}
+	return ""
 }
 
 type Code struct {
@@ -138,19 +147,71 @@ func (x *Code) GetCode() string {
 	return ""
 }
 
+type Stats struct {
+	state         protoimpl.MessageState
+	sizeCache     protoimpl.SizeCache
+	unknownFields protoimpl.UnknownFields
+
+	Stats string `protobuf:"bytes,1,opt,name=Stats,proto3" json:"Stats,omitempty"`
+}
+
+func (x *Stats) Reset() {
+	*x = Stats{}
+	if protoimpl.UnsafeEnabled {
+		mi := &file_url_service_proto_msgTypes[2]
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		ms.StoreMessageInfo(mi)
+	}
+}
+
+func (x *Stats) String() string {
+	return protoimpl.X.MessageStringOf(x)
+}
+
+func (*Stats) ProtoMessage() {}
+
+func (x *Stats) ProtoReflect() protoreflect.Message {
+	mi := &file_url_service_proto_msgTypes[2]
+	if protoimpl.UnsafeEnabled && x != nil {
+		ms := protoimpl.X.MessageStateOf(protoimpl.Pointer(x))
+		if ms.LoadMessageInfo() == nil {
+			ms.StoreMessageInfo(mi)
+		}
+		return ms
+	}
+	return mi.MessageOf(x)
+}
+
+// Deprecated: Use Stats.ProtoReflect.Descriptor instead.
+func (*Stats) Descriptor() ([]byte, []int) {
+	return file_url_service_proto_rawDescGZIP(), []int{2}
+}
+
+func (x *Stats) GetStats() string {
+	if x != nil {
+		return x.Stats
+	}
+	return ""
+}
+
 var File_url_service_proto protoreflect.FileDescriptor
 
 var file_url_service_proto_rawDesc = []byte{
 	0x0a, 0x11, 0x75, 0x72, 0x6c, 0x5f, 0x73, 0x65, 0x72, 0x76, 0x69, 0x63, 0x65, 0x2e, 0x70, 0x72,
-	0x6f, 0x74, 0x6f, 0x22, 0x31, 0x0a, 0x03, 0x55, 0x52, 0x4c, 0x12, 0x18, 0x0a, 0x07, 0x4c, 0x6f,
+	0x6f, 0x74, 0x6f, 0x22, 0x47, 0x0a, 0x03, 0x55, 0x52, 0x4c, 0x12, 0x18, 0x0a, 0x07, 0x4c, 0x6f,
 	0x6e, 0x67, 0x55, 0x72, 0x6c, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x07, 0x4c, 0x6f, 0x6e,
 	0x67, 0x55, 0x72, 0x6c, 0x12, 0x10, 0x0a, 0x03, 0x54, 0x54, 0x4c, 0x18, 0x02, 0x20, 0x01, 0x28,
-	0x03, 0x52, 0x03, 0x54, 0x54, 0x4c, 0x22, 0x1a, 0x0a, 0x04, 0x43, 0x6f, 0x64, 0x65, 0x12, 0x12,
-	0x0a, 0x04, 0x43, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09, 0x52, 0x04, 0x43, 0x6f,
-	0x64, 0x65, 0x32, 0x3c, 0x0a, 0x09, 0x53, 0x68, 0x6f, 0x72, 0x74, 0x65, 0x6e, 0x65, 0x72, 0x12,
-	0x18, 0x0a, 0x09, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x55, 0x52, 0x4c, 0x12, 0x04, 0x2e, 0x55,
-	0x52, 0x4c, 0x1a, 0x05, 0x2e, 0x43, 0x6f, 0x64, 0x65, 0x12, 0x15, 0x0a, 0x06, 0x47, 0x65, 0x74,
-	0x55, 0x52, 0x4c, 0x12, 0x05, 0x2e, 0x43, 0x6f, 0x64, 0x65, 0x1a, 0x04, 0x2e, 0x55, 0x52, 0x4c,
+	0x03, 0x52, 0x03, 0x54, 0x54, 0x4c, 0x12, 0x14, 0x0a, 0x05, 0x53, 0x74, 0x61, 0x74, 0x73, 0x18,
+	0x03, 0x20, 0x01, 0x28, 0x09, 0x52, 0x05, 0x53, 0x74, 0x61, 0x74, 0x73, 0x22, 0x1a, 0x0a, 0x04,
+	0x43, 0x6f, 0x64, 0x65, 0x12, 0x12, 0x0a, 0x04, 0x43, 0x6f, 0x64, 0x65, 0x18, 0x01, 0x20, 0x01,
+	0x28, 0x09, 0x52, 0x04, 0x43, 0x6f, 0x64, 0x65, 0x22, 0x1d, 0x0a, 0x05, 0x53, 0x74, 0x61, 0x74,
+	0x73, 0x12, 0x14, 0x0a, 0x05, 0x53, 0x74, 0x61, 0x74, 0x73, 0x18, 0x01, 0x20, 0x01, 0x28, 0x09,
+	0x52, 0x05, 0x53, 0x74, 0x61, 0x74, 0x73, 0x32, 0x57, 0x0a, 0x09, 0x53, 0x68, 0x6f, 0x72, 0x74,
+	0x65, 0x6e, 0x65, 0x72, 0x12, 0x18, 0x0a, 0x09, 0x43, 0x72, 0x65, 0x61, 0x74, 0x65, 0x55, 0x52,
+	0x4c, 0x12, 0x04, 0x2e, 0x55, 0x52, 0x4c, 0x1a, 0x05, 0x2e, 0x43, 0x6f, 0x64, 0x65, 0x12, 0x15,
+	0x0a, 0x06, 0x47, 0x65, 0x74, 0x55, 0x52, 0x4c, 0x12, 0x05, 0x2e, 0x43, 0x6f, 0x64, 0x65, 0x1a,
+	0x04, 0x2e, 0x55, 0x52, 0x4c, 0x12, 0x19, 0x0a, 0x08, 0x47, 0x65, 0x74, 0x53, 0x74, 0x61, 0x74,
+	0x73, 0x12, 0x05, 0x2e, 0x43, 0x6f, 0x64, 0x65, 0x1a, 0x06, 0x2e, 0x53, 0x74, 0x61, 0x74, 0x73,
 	0x42, 0x08, 0x5a, 0x06, 0x70, 0x72, 0x6f, 0x74, 0x6f, 0x73, 0x62, 0x06, 0x70, 0x72, 0x6f, 0x74,
 	0x6f, 0x33,
 }
@@ -167,18 +228,21 @@ func file_url_service_proto_rawDescGZIP() []byte {
 	return file_url_service_proto_rawDescData
 }
 
-var file_url_service_proto_msgTypes = make([]protoimpl.MessageInfo, 2)
+var file_url_service_proto_msgTypes = make([]protoimpl.MessageInfo, 3)
 var file_url_service_proto_goTypes = []interface{}{
-	(*URL)(nil),  // 0: URL
-	(*Code)(nil), // 1: Code
+	(*URL)(nil),   // 0: URL
+	(*Code)(nil),  // 1: Code
+	(*Stats)(nil), // 2: Stats
 }
 var file_url_service_proto_depIdxs = []int32{
 	0, // 0: Shortener.CreateURL:input_type -> URL
 	1, // 1: Shortener.GetURL:input_type -> Code
-	1, // 2: Shortener.CreateURL:output_type -> Code
-	0, // 3: Shortener.GetURL:output_type -> URL
-	2, // [2:4] is the sub-list for method output_type
-	0, // [0:2] is the sub-list for method input_type
+	1, // 2: Shortener.GetStats:input_type -> Code
+	1, // 3: Shortener.CreateURL:output_type -> Code
+	0, // 4: Shortener.GetURL:output_type -> URL
+	2, // 5: Shortener.GetStats:output_type -> Stats
+	3, // [3:6] is the sub-list for method output_type
+	0, // [0:3] is the sub-list for method input_type
 	0, // [0:0] is the sub-list for extension type_name
 	0, // [0:0] is the sub-list for extension extendee
 	0, // [0:0] is the sub-list for field type_name
@@ -214,6 +278,18 @@ func file_url_service_proto_init() {
 				return nil
 			}
 		}
+		file_url_service_proto_msgTypes[2].Exporter = func(v interface{}, i int) interface{} {
+			switch v := v.(*Stats); i {
+			case 0:
+				return &v.state
+			case 1:
+				return &v.sizeCache
+			case 2:
+				return &v.unknownFields
+			default:
+				return nil
+			}
+		}
 	}
 	type x struct{}
 	out := protoimpl.TypeBuilder{
@@ -221,7 +297,7 @@ func file_url_service_proto_init() {
 			GoPackagePath: reflect.TypeOf(x{}).PkgPath(),
 			RawDescriptor: file_url_service_proto_rawDesc,
 			NumEnums:      0,
-			NumMessages:   2,
+			NumMessages:   3,
 			NumExtensions: 0,
 			NumServices:   1,
 		},
@@ -249,6 +325,7 @@ const _ = grpc.SupportPackageIsVersion6
 type ShortenerClient interface {
 	CreateURL(ctx context.Context, in *URL, opts ...grpc.CallOption) (*Code, error)
 	GetURL(ctx context.Context, in *Code, opts ...grpc.CallOption) (*URL, error)
+	GetStats(ctx context.Context, in *Code, opts ...grpc.CallOption) (*Stats, error)
 }
 
 type shortenerClient struct {
@@ -277,10 +354,20 @@ func (c *shortenerClient) GetURL(ctx context.Context, in *Code, opts ...grpc.Cal
 	return out, nil
 }
 
+func (c *shortenerClient) GetStats(ctx context.Context, in *Code, opts ...grpc.CallOption) (*Stats, error) {
+	out := new(Stats)
+	err := c.cc.Invoke(ctx, "/Shortener/GetStats", in, out, opts...)
+	if err != nil {
+		return nil, err
+	}
+	return out, nil
+}
+
 // ShortenerServer is the server API for Shortener service.
 type ShortenerServer interface {
 	CreateURL(context.Context, *URL) (*Code, error)
 	GetURL(context.Context, *Code) (*URL, error)
+	GetStats(context.Context, *Code) (*Stats, error)
 }
 
 // UnimplementedShortenerServer can be embedded to have forward compatible implementations.
@@ -292,6 +379,9 @@ func (*UnimplementedShortenerServer) CreateURL(context.Context, *URL) (*Code, er
 }
 func (*UnimplementedShortenerServer) GetURL(context.Context, *Code) (*URL, error) {
 	return nil, status.Errorf(codes.Unimplemented, "method GetURL not implemented")
+}
+func (*UnimplementedShortenerServer) GetStats(context.Context, *Code) (*Stats, error) {
+	return nil, status.Errorf(codes.Unimplemented, "method GetStats not implemented")
 }
 
 func RegisterShortenerServer(s *grpc.Server, srv ShortenerServer) {
@@ -334,6 +424,24 @@ func _Shortener_GetURL_Handler(srv interface{}, ctx context.Context, dec func(in
 	return interceptor(ctx, in, info, handler)
 }
 
+func _Shortener_GetStats_Handler(srv interface{}, ctx context.Context, dec func(interface{}) error, interceptor grpc.UnaryServerInterceptor) (interface{}, error) {
+	in := new(Code)
+	if err := dec(in); err != nil {
+		return nil, err
+	}
+	if interceptor == nil {
+		return srv.(ShortenerServer).GetStats(ctx, in)
+	}
+	info := &grpc.UnaryServerInfo{
+		Server:     srv,
+		FullMethod: "/Shortener/GetStats",
+	}
+	handler := func(ctx context.Context, req interface{}) (interface{}, error) {
+		return srv.(ShortenerServer).GetStats(ctx, req.(*Code))
+	}
+	return interceptor(ctx, in, info, handler)
+}
+
 var _Shortener_serviceDesc = grpc.ServiceDesc{
 	ServiceName: "Shortener",
 	HandlerType: (*ShortenerServer)(nil),
@@ -345,6 +453,10 @@ var _Shortener_serviceDesc = grpc.ServiceDesc{
 		{
 			MethodName: "GetURL",
 			Handler:    _Shortener_GetURL_Handler,
+		},
+		{
+			MethodName: "GetStats",
+			Handler:    _Shortener_GetStats_Handler,
 		},
 	},
 	Streams:  []grpc.StreamDesc{},